refactor(webhook): extract bubble generation so that we can reuse them in batch processing
