+++ src/webhook/handlers/initState.ts
@@ -1,28 +1,23 @@
-import stringSimilarity from 'string-similarity';
 import { t } from 'ttag';
-import {
-  FlexBubble,
-  FlexComponent,
-  FlexMessage,
-  Message,
-  TextMessage,
-} from '@line/bot-sdk';
+import { FlexBubble, FlexMessage, Message, TextMessage } from '@line/bot-sdk';
+
 import type {
   Result,
   Context,
   CooccurredMessage,
 } from 'src/types/chatbotState';
+import ga from 'src/lib/ga';
+import detectDialogflowIntent from 'src/lib/detectDialogflowIntent';
+
 import {
   createPostbackAction,
   ellipsis,
   POSTBACK_NO_ARTICLE_FOUND,
-  createHighlightContents,
   createTextMessage,
   createArticleSourceReply,
   searchText,
+  createTextCarouselContents,
 } from './utils';
-import ga from 'src/lib/ga';
-import detectDialogflowIntent from 'src/lib/detectDialogflowIntent';
 import choosingArticle from './choosingArticle';
 
 const SIMILARITY_THRESHOLD = 0.95;
@@ -81,7 +76,7 @@ const initState = async ({
 
   const inputSummary = ellipsis(input, 12);
 
-  if (result?.edges.length) {
+  if (result.edges.length) {
     // Track if find similar Articles in DB.
     visitor.event({ ec: 'UserInput', ea: 'ArticleSearch', el: 'ArticleFound' });
 
@@ -95,23 +90,9 @@ const initState = async ({
       });
     });
 
-    const edgesSortedWithSimilarity = result.edges
-      .map((edge) => ({
-        ...edge,
-        similarity: stringSimilarity.compareTwoStrings(
-          // Remove spaces so that we count word's similarities only
-          //
-          (edge.node.text ?? '').replace(/\s/g, ''),
-          input.replace(/\s/g, '')
-        ),
-      }))
-      .sort((edge1, edge2) => edge2.similarity - edge1.similarity)
-      .slice(0, 9); /* flex carousel has at most 10 bubbles */
+    const hasIdenticalDocs = result.edges[0].similarity >= SIMILARITY_THRESHOLD;
 
-    const hasIdenticalDocs =
-      edgesSortedWithSimilarity[0].similarity >= SIMILARITY_THRESHOLD;
-
-    if (edgesSortedWithSimilarity.length === 1 && hasIdenticalDocs) {
+    if (result.edges.length === 1 && hasIdenticalDocs) {
       visitor.send();
 
       return await choosingArticle({
@@ -120,107 +101,15 @@ const initState = async ({
         postbackData: {
           sessionId: context.sessionId,
           state: 'CHOOSING_ARTICLE',
-          input: edgesSortedWithSimilarity[0].node.id,
+          input: result.edges[0].node.id,
         },
         userId,
       });
     }
 
-    const articleOptions: FlexBubble[] = edgesSortedWithSimilarity.map(
-      ({ node: { text, id, articleType }, highlight, similarity }) => {
-        const similarityPercentage = Math.round(similarity * 100);
-        const similarityEmoji = ['😐', '🙂', '😀', '😃', '😄'][
-          Math.floor(similarity * 4.999)
-        ];
-        const displayTextWhenChosen = ellipsis(text ?? '', 25, '...');
-
-        const bodyContents: FlexComponent[] = [];
-
-        const { contents: highlightContents, source: highlightSource } =
-          createHighlightContents(highlight);
-
-        let highlightSourceInfo = '';
-        switch (highlightSource) {
-          case 'hyperlinks':
-            highlightSourceInfo = t`(Words found in the hyperlink)`;
-            break;
-          case 'text':
-            if (articleType !== 'TEXT') {
-              highlightSourceInfo = t`(Words found in transcript)`;
-            }
-        }
-        if (highlightSourceInfo) {
-          bodyContents.push({
-            type: 'text',
-            text: highlightSourceInfo,
-            size: 'sm',
-            color: '#ff7b7b',
-            weight: 'bold',
-          });
-        }
-
-        bodyContents.push({
-          type: 'text',
-          contents: highlightContents,
-          maxLines: 6,
-          flex: 0,
-          gravity: 'top',
-          weight: 'regular',
-          wrap: true,
-        });
-
-        return {
-          type: 'bubble',
-          direction: 'ltr',
-          header: {
-            type: 'box',
-            layout: 'horizontal',
-            spacing: 'md',
-            paddingBottom: 'none',
-            contents: [
-              {
-                type: 'text',
-                text: similarityEmoji,
-                flex: 0,
-              },
-              {
-                type: 'text',
-                text: t`Looks ${similarityPercentage}% similar`,
-                gravity: 'center',
-                size: 'sm',
-                weight: 'bold',
-                wrap: true,
-                color: '#AAAAAA',
-              },
-            ],
-          },
-          body: {
-            type: 'box',
-            layout: 'vertical',
-            spacing: 'none',
-            margin: 'none',
-            contents: bodyContents,
-          },
-          footer: {
-            type: 'box',
-            layout: 'horizontal',
-            contents: [
-              {
-                type: 'button',
-                action: createPostbackAction(
-                  t`Choose this one`,
-                  id,
-                  t`I choose “${displayTextWhenChosen}”`,
-                  context.sessionId,
-                  'CHOOSING_ARTICLE'
-                ),
-                style: 'primary',
-                color: '#ffb600',
-              },
-            ],
-          },
-        };
-      }
+    const articleOptions: FlexBubble[] = createTextCarouselContents(
+      result.edges,
+      context.sessionId
     );
 
     // Show "no-article-found" option only when no identical docs are found
