+++ src/webhook/handlers/utils.ts
@@ -11,6 +11,7 @@ import type {
 } from '@line/bot-sdk';
 import { t, msgid, ngettext } from 'ttag';
 import GraphemeSplitter from 'grapheme-splitter';
+import stringSimilarity from 'string-similarity';
 
 import gql from 'src/lib/gql';
 import { getArticleURL, createTypeWords, format } from 'src/lib/sharedUtils';
@@ -864,9 +865,26 @@ export function getLineContentProxyURL(messageId: string) {
   return `${process.env.RUMORS_LINE_BOT_URL}/getcontent?token=${jwt}`;
 }
 
-export async function searchText(
-  text: string
-): Promise<ListArticlesInInitStateQuery['ListArticles']> {
+/**
+ * ListArticle result with similarity score
+ */
+type SearchTextResult = Omit<
+  ListArticlesInInitStateQuery['ListArticles'],
+  'edges'
+> & {
+  edges: Array<
+    NonNullable<
+      ListArticlesInInitStateQuery['ListArticles']
+    >['edges'][number] & {
+      similarity: number;
+    }
+  >;
+};
+
+/**
+ * Searches for text and reorder with string similarity
+ */
+export async function searchText(text: string): Promise<SearchTextResult> {
   const {
     data: { ListArticles },
   } = await gql`
@@ -895,13 +913,142 @@ export async function searchText(
   `<ListArticlesInInitStateQuery, ListArticlesInInitStateQueryVariables>({
     text,
   });
-  return ListArticles;
+
+  const sanitizedText = text.replace(/\s/g, '');
+  const edgesSortedWithSimilarity =
+    ListArticles?.edges
+      .map((edge) => ({
+        ...edge,
+        similarity: stringSimilarity.compareTwoStrings(
+          // Remove spaces so that we count word's similarities only
+          //
+          (edge.node.text ?? '').replace(/\s/g, ''),
+          sanitizedText
+        ),
+      }))
+      .sort((edge1, edge2) => edge2.similarity - edge1.similarity) ?? [];
+
+  return {
+    ...ListArticles,
+    edges: edgesSortedWithSimilarity,
+  };
 }
+
+export function createTextCarouselContents(
+  edges: SearchTextResult['edges'],
+  sessionId: number
+) {
+  return edges
+    .map<FlexBubble>(
+      ({ node: { text, id, articleType }, highlight, similarity }) => {
+        const similarityPercentage = Math.round(similarity * 100);
+        const similarityEmoji = ['😐', '🙂', '😀', '😃', '😄'][
+          Math.floor(similarity * 4.999)
+        ];
+        const displayTextWhenChosen = ellipsis(text ?? '', 25, '...');
+
+        const bodyContents: FlexComponent[] = [];
+
+        const { contents: highlightContents, source: highlightSource } =
+          createHighlightContents(highlight);
+
+        let highlightSourceInfo = '';
+        switch (highlightSource) {
+          case 'hyperlinks':
+            highlightSourceInfo = t`(Words found in the hyperlink)`;
+            break;
+          case 'text':
+            if (articleType !== 'TEXT') {
+              highlightSourceInfo = t`(Words found in transcript)`;
+            }
+        }
+        if (highlightSourceInfo) {
+          bodyContents.push({
+            type: 'text',
+            text: highlightSourceInfo,
+            size: 'sm',
+            color: '#ff7b7b',
+            weight: 'bold',
+          });
+        }
+
+        bodyContents.push({
+          type: 'text',
+          contents: highlightContents,
+          maxLines: 6,
+          flex: 0,
+          gravity: 'top',
+          weight: 'regular',
+          wrap: true,
+        });
+
+        return {
+          type: 'bubble',
+          direction: 'ltr',
+          header: {
+            type: 'box',
+            layout: 'horizontal',
+            spacing: 'md',
+            paddingBottom: 'none',
+            contents: [
+              {
+                type: 'text',
+                text: similarityEmoji,
+                flex: 0,
+              },
+              {
+                type: 'text',
+                text: t`Looks ${similarityPercentage}% similar`,
+                gravity: 'center',
+                size: 'sm',
+                weight: 'bold',
+                wrap: true,
+                color: '#AAAAAA',
+              },
+            ],
+          },
+          body: {
+            type: 'box',
+            layout: 'vertical',
+            spacing: 'none',
+            margin: 'none',
+            contents: bodyContents,
+          },
+          footer: {
+            type: 'box',
+            layout: 'horizontal',
+            contents: [
+              {
+                type: 'button',
+                action: createPostbackAction(
+                  t`Choose this one`,
+                  id,
+                  t`I choose “${displayTextWhenChosen}”`,
+                  sessionId,
+                  'CHOOSING_ARTICLE'
+                ),
+                style: 'primary',
+                color: '#ffb600',
+              },
+            ],
+          },
+        };
+      }
+    ) /* flex carousel has at most 10 bubbles */
+    .slice(0, 9);
+}
+
+type SearchMediaResult = Omit<
+  ListArticlesInProcessMediaQuery['ListArticles'],
+  'edges'
+> & {
+  edges: NonNullable<ListArticlesInProcessMediaQuery['ListArticles']>['edges'];
+};
 
 export async function searchMedia(
   mediaUrl: string,
   userId: string
-): Promise<ListArticlesInProcessMediaQuery['ListArticles']> {
+): Promise<SearchMediaResult> {
   const {
     data: { ListArticles },
   } = await gql`
@@ -937,5 +1084,145 @@ export async function searchMedia(
     { mediaUrl },
     { userId }
   );
-  return ListArticles;
+  return {
+    ...ListArticles,
+    edges: [...(ListArticles?.edges ?? [])].sort(
+      (a, b) => b.mediaSimilarity - a.mediaSimilarity
+    ),
+  };
+}
+
+const CIRCLED_DIGITS = '⓪①②③④⑤⑥⑦⑧⑨⑩⑪';
+
+/**
+ * @param edges - edge data returned by searchMedia()
+ * @param sessionId
+ * @returns
+ */
+export function createMediaCarouselContents(
+  edges: SearchMediaResult['edges'],
+  sessionId: number
+): FlexBubble[] {
+  return edges
+    .map(
+      (
+        {
+          node: { attachmentUrl, id, articleType },
+          highlight,
+          mediaSimilarity,
+        },
+        index
+      ): FlexBubble => {
+        const displayTextWhenChosen = CIRCLED_DIGITS[index + 1];
+
+        const { contents: highlightContents, source: highlightSource } =
+          createHighlightContents(highlight);
+
+        const similarityPercentage = Math.round(mediaSimilarity * 100);
+
+        const looks =
+          mediaSimilarity > 0
+            ? t`Looks ${similarityPercentage}% similar`
+            : highlightSource === null
+            ? t`Similar file`
+            : t`Contains relevant text`;
+
+        const bodyContents: FlexComponent[] = [];
+
+        if (highlightSource) {
+          let highlightSourceInfo = '';
+          switch (highlightSource) {
+            case 'hyperlinks':
+              highlightSourceInfo = t`(Text in the hyperlink)`;
+              break;
+            case 'text':
+              if (articleType !== 'TEXT') {
+                highlightSourceInfo = t`(Text in transcript)`;
+              }
+          }
+
+          if (highlightSourceInfo) {
+            bodyContents.push({
+              type: 'text',
+              text: highlightSourceInfo,
+              size: 'sm',
+              color: '#ff7b7b',
+              weight: 'bold',
+            });
+          }
+
+          bodyContents.push({
+            type: 'text',
+            contents: highlightContents,
+            // Show less lines if there are thumbnails to show
+            maxLines: attachmentUrl ? 5 : 12,
+            flex: 0,
+            gravity: 'top',
+            weight: 'regular',
+            wrap: true,
+          });
+        }
+
+        return {
+          type: 'bubble',
+          direction: 'ltr',
+          header: {
+            type: 'box',
+            layout: 'horizontal',
+            spacing: 'sm',
+            paddingBottom: 'md',
+            contents: [
+              {
+                type: 'text',
+                text: displayTextWhenChosen + ' ' + looks,
+                gravity: 'center',
+                size: 'sm',
+                weight: 'bold',
+                wrap: true,
+                color: '#AAAAAA',
+              },
+            ],
+          },
+
+          // Show thumbnail image if available
+          hero: !attachmentUrl
+            ? undefined
+            : {
+                type: 'image',
+                url: attachmentUrl,
+                size: 'full',
+              },
+
+          // Show highlighted text if available
+          body:
+            bodyContents.length === 0
+              ? undefined
+              : {
+                  type: 'box',
+                  layout: 'vertical',
+                  contents: bodyContents,
+                },
+
+          footer: {
+            type: 'box',
+            layout: 'horizontal',
+            contents: [
+              {
+                type: 'button',
+                action: createPostbackAction(
+                  t`Choose this one`,
+                  id,
+                  t`I choose ${displayTextWhenChosen}`,
+                  sessionId,
+                  'CHOOSING_ARTICLE'
+                ),
+                style: 'primary',
+                color: '#ffb600',
+              },
+            ],
+          },
+        };
+      }
+    )
+    .slice(0, 9); /* flex carousel has at most 10 bubbles */
 }
